feat(auth): [BACK-1345] add auth to getApprovedCuratedCorpusItemByUrl query
-move auth tests to a separate file
