+++ src/admin/resolvers/queries/ApprovedItem.integration.ts
@@ -11,7 +11,7 @@ import {
   GET_APPROVED_ITEMS,
 } from './sample-queries.gql';
 import { CuratedCorpusEventEmitter } from '../../../events/curatedCorpusEventEmitter';
-import { ACCESS_DENIED_ERROR, MozillaAccessGroup } from '../../../shared/types';
+import { MozillaAccessGroup } from '../../../shared/types';
 
 describe('queries: ApprovedCuratedCorpusItem', () => {
   // adding headers with groups that grant full access
@@ -315,7 +315,7 @@ describe('queries: ApprovedCuratedCorpusItem', () => {
     });
   });
 
-  describe('getApprovedCuratedCorpusItems query', () => {
+  describe('getApprovedCuratedCorpusItemByUrl query', () => {
     beforeAll(async () => {
       // Create a few items with known URLs.
       const storyInput = [
@@ -387,140 +387,5 @@ describe('queries: ApprovedCuratedCorpusItem', () => {
       // There should be no errors
       expect(result.errors).to.be.undefined;
     });
-  });
-});
-
-describe('queries: ApprovedCuratedCorpusItem - authentication checks', () => {
-  beforeAll(async () => {
-    // clear out db to start fresh
-    await clearDb(db);
-
-    // Create some items
-    const stories = [
-      {
-        title: 'How To Win Friends And Influence People with GraphQL',
-        language: 'en',
-        status: CuratedStatus.RECOMMENDATION,
-        topic: 'FOOD',
-      },
-      {
-        title: 'What Zombies Can Teach You About GraphQL',
-        language: 'en',
-        status: CuratedStatus.RECOMMENDATION,
-        url: 'https://www.sample-domain/what-zombies-can-teach-you-graphql',
-        topic: 'TECHNOLOGY',
-      },
-      {
-        title: 'How To Make Your Product Stand Out With GraphQL',
-        language: 'en',
-        status: CuratedStatus.RECOMMENDATION,
-        topic: 'TECHNOLOGY',
-      },
-    ];
-
-    // insert test stories into the db
-    for (const story of stories) {
-      await createApprovedItemHelper(db, story);
-    }
-  });
-
-  afterAll(async () => {
-    await db.$disconnect();
-  });
-
-  it('should get all items when user has read-only access', async () => {
-    // Set up auth headers with read-only access
-    const headers = {
-      name: 'Test User',
-      username: 'test.user@test.com',
-      groups: `group1,group2,${MozillaAccessGroup.READONLY}`,
-    };
-
-    const server = getServerWithMockedHeaders(headers);
-    await server.start();
-
-    const result = await server.executeOperation({
-      query: GET_APPROVED_ITEMS,
-    });
-
-    expect(result.errors).to.be.undefined;
-    expect(result.data).not.to.be.null;
-
-    // should return all 3 items
-    expect(result.data?.getApprovedCuratedCorpusItems.edges).to.have.length(3);
-
-    await server.stop();
-  });
-
-  it('should get all items when user has only one scheduled surface access', async () => {
-    // Set up auth headers with access to a single Scheduled Surface
-    const headers = {
-      name: 'Test User',
-      username: 'test.user@test.com',
-      groups: `group1,group2,${MozillaAccessGroup.NEW_TAB_CURATOR_ENUS}`,
-    };
-
-    const server = getServerWithMockedHeaders(headers);
-    await server.start();
-
-    const result = await server.executeOperation({
-      query: GET_APPROVED_ITEMS,
-    });
-
-    expect(result.errors).to.be.undefined;
-    expect(result.data).not.to.be.null;
-
-    // should return all 3 items
-    expect(result.data?.getApprovedCuratedCorpusItems.edges).to.have.length(3);
-
-    await server.stop();
-  });
-
-  it('should throw an error when user does not have the required access', async () => {
-    // Set up auth headers with no valid access group
-    const headers = {
-      name: 'Test User',
-      username: 'test.user@test.com',
-      groups: `group1,group2`,
-    };
-
-    const server = getServerWithMockedHeaders(headers);
-    await server.start();
-
-    const result = await server.executeOperation({
-      query: GET_APPROVED_ITEMS,
-    });
-
-    expect(result.data).to.be.null;
-    expect(result.errors).not.to.be.undefined;
-
-    // check if the error we get is access denied error
-    expect(result.errors?.[0].message).to.equal(ACCESS_DENIED_ERROR);
-
-    await server.stop();
-  });
-
-  it('should throw an error when request access groups are undefined', async () => {
-    // Set up auth headers with no valid access group
-    const headers = {
-      name: 'Test User',
-      username: 'test.user@test.com',
-      groups: undefined,
-    };
-
-    const server = getServerWithMockedHeaders(headers);
-    await server.start();
-
-    const result = await server.executeOperation({
-      query: GET_APPROVED_ITEMS,
-    });
-
-    expect(result.data).to.be.null;
-    expect(result.errors).not.to.be.undefined;
-
-    // check if the error we get is access denied error
-    expect(result.errors?.[0].message).to.equal(ACCESS_DENIED_ERROR);
-
-    await server.stop();
   });
 });
